Extract RepositoryPaginator into its own module
