+++ src/components/mod.rs
@@ -3,3 +3,4 @@ pub mod about;
 pub mod organization_entry;
 pub mod repository_card;
 pub mod repository_list;
+pub mod repository_paginator;
